pyproject.toml
[tool.poetry]
name = "housetrack"
version = "0.1.0"
description = ""
authors = ["joostsijm <joostsijm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyairtable = "^1.5.0"
python-dotenv = "^1.0.0"
selenium = "^4.10.0"
beautifulsoup4 = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
